docs: Update changelog regarding license change and bump version
This commit updates the `README.md` to announce a future license change.

Specifically, it adds a note to the changelog stating that version `1.1.78` is the last version to be released under the MIT License. Starting from version `1.1.79`, the project will be licensed under Apache 2.0.

Additionally, the project version in `gradle.properties` has been incremented from `1.1.77` to `1.1.78`.